@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import os
+import sys
 
 try:
     from setuptools import setup
@@ -37,16 +38,28 @@ AUTHOR = 'pvlib python Developers'
 MAINTAINER_EMAIL = 'holmgren@email.arizona.edu'
 URL = 'https://github.com/pvlib/pvlib-python'
 
-INSTALL_REQUIRES = ['numpy >= 1.10.4',
-                    'pandas >= 0.18.1',
+INSTALL_REQUIRES = ['numpy >= 1.12.0',
+                    'pandas >= 0.22.0',
                     'pytz',
-                    'requests']
+                    'requests',
+                    'scipy >= 1.2.0',
+                    'h5py']
+
+# include dataclasses as a dependency only on python 3.6
+if sys.version_info.major == 3 and sys.version_info.minor == 6:
+    INSTALL_REQUIRES.append('dataclasses')
+
 TESTS_REQUIRE = ['nose', 'pytest', 'pytest-cov', 'pytest-mock',
-                 'pytest-timeout']
+                 'requests-mock', 'pytest-timeout', 'pytest-rerunfailures',
+                 'pytest-remotedata']
 EXTRAS_REQUIRE = {
-    'optional': ['ephem', 'cython', 'netcdf4', 'nrel-pysam', 'numba',
-                 'pvfactors', 'scipy', 'siphon', 'tables'],
-    'doc': ['ipython', 'matplotlib', 'sphinx', 'sphinx_rtd_theme'],
+    'optional': ['cython', 'ephem', 'netcdf4', 'nrel-pysam', 'numba',
+                 'pvfactors', 'siphon', 'statsmodels',
+                 'cftime >= 1.1.1'],
+    'doc': ['ipython', 'matplotlib', 'sphinx == 3.1.2',
+            'sphinx_rtd_theme==0.5.0', 'sphinx-gallery', 'docutils == 0.15.2',
+            'pillow', 'netcdf4', 'siphon',
+            'sphinx-toggleprompt >= 0.0.5'],
     'test': TESTS_REQUIRE
 }
 EXTRAS_REQUIRE['all'] = sorted(set(sum(EXTRAS_REQUIRE.values(), [])))
@@ -58,9 +71,6 @@ CLASSIFIERS = [
     'Intended Audience :: Science/Research',
     'Programming Language :: Python',
     'Programming Language :: Python :: 3',
-    'Programming Language :: Python :: 3.5',
-    'Programming Language :: Python :: 3.6',
-    'Programming Language :: Python :: 3.7',
     'Topic :: Scientific/Engineering',
 ]
 
@@ -68,7 +78,13 @@ setuptools_kwargs = {
     'zip_safe': False,
     'scripts': [],
     'include_package_data': True,
-    'python_requires': '~=3.5'
+    'python_requires': '>=3.6'
+}
+
+PROJECT_URLS = {
+    "Bug Tracker": "https://github.com/pvlib/pvlib-python/issues",
+    "Documentation": "https://pvlib-python.readthedocs.io/",
+    "Source Code": "https://github.com/pvlib/pvlib-python",
 }
 
 # set up pvlib packages to be installed and extensions to be compiled
@@ -107,5 +123,6 @@ setup(name=DISTNAME,
       maintainer_email=MAINTAINER_EMAIL,
       license=LICENSE,
       url=URL,
+      project_urls=PROJECT_URLS,
       classifiers=CLASSIFIERS,
       **setuptools_kwargs)
