@@ -22,12 +22,12 @@ def pres2alt(pressure):
     Parameters
     ----------
     pressure : numeric
-        Atmospheric pressure (Pascals)
+        Atmospheric pressure. [Pa]
 
     Returns
     -------
     altitude : numeric
-        Altitude in meters above sea level
+        Altitude above sea level. [m]
 
     Notes
     ------
@@ -40,14 +40,14 @@ def pres2alt(pressure):
     Temperature at zero altitude   288.15 K
     Gravitational acceleration     9.80665 m/s^2
     Lapse rate                     -6.5E-3 K/m
-    Gas constant for air           287.053 J/(kgK)
+    Gas constant for air           287.053 J/(kg K)
     Relative Humidity              0%
     ============================   ================
 
     References
     -----------
-    [1] "A Quick Derivation relating altitude to air pressure" from
-    Portland State Aerospace Society, Version 1.03, 12/22/2004.
+    .. [1] "A Quick Derivation relating altitude to air pressure" from
+       Portland State Aerospace Society, Version 1.03, 12/22/2004.
     '''
 
     alt = 44331.5 - 4946.62 * pressure ** (0.190263)
@@ -62,12 +62,12 @@ def alt2pres(altitude):
     Parameters
     ----------
     altitude : numeric
-        Altitude in meters above sea level
+        Altitude above sea level. [m]
 
     Returns
     -------
     pressure : numeric
-        Atmospheric pressure (Pascals)
+        Atmospheric pressure. [Pa]
 
     Notes
     ------
@@ -80,14 +80,14 @@ def alt2pres(altitude):
     Temperature at zero altitude   288.15 K
     Gravitational acceleration     9.80665 m/s^2
     Lapse rate                     -6.5E-3 K/m
-    Gas constant for air           287.053 J/(kgK)
+    Gas constant for air           287.053 J/(kg K)
     Relative Humidity              0%
     ============================   ================
 
     References
     -----------
-    [1] "A Quick Derivation relating altitude to air pressure" from
-    Portland State Aerospace Society, Version 1.03, 12/22/2004.
+    .. [1] "A Quick Derivation relating altitude to air pressure" from
+       Portland State Aerospace Society, Version 1.03, 12/22/2004.
     '''
 
     press = 100 * ((44331.514 - altitude) / 11880.516) ** (1 / 0.1902632)
@@ -96,37 +96,36 @@ def alt2pres(altitude):
 
 
 def get_absolute_airmass(airmass_relative, pressure=101325.):
-    '''
-    Determine absolute (pressure corrected) airmass from relative
-    airmass and pressure
+    r'''
+    Determine absolute (pressure-adjusted) airmass from relative
+    airmass and pressure.
 
-    Gives the airmass for locations not at sea-level (i.e. not at
-    standard pressure). The input argument "AMrelative" is the relative
-    airmass. The input argument "pressure" is the pressure (in Pascals)
-    at the location of interest and must be greater than 0. The
-    calculation for absolute airmass is
+    The calculation for absolute airmass (:math:`AM_a`) is
 
     .. math::
-        absolute airmass = (relative airmass)*pressure/101325
+        AM_a = AM_r \frac{P}{101325}
+
+    where :math:`AM_r` is relative air mass at sea level and :math:`P` is
+    atmospheric pressure.
 
     Parameters
     ----------
     airmass_relative : numeric
-        The airmass at sea-level.
+        The airmass at sea level. [unitless]
 
     pressure : numeric, default 101325
-        The site pressure in Pascal.
+        Atmospheric pressure. [Pa]
 
     Returns
     -------
     airmass_absolute : numeric
-        Absolute (pressure corrected) airmass
+        Absolute (pressure-adjusted) airmass
 
     References
     ----------
-    [1] C. Gueymard, "Critical analysis and performance assessment of
-    clear sky solar irradiance models using theoretical and measured
-    data," Solar Energy, vol. 51, pp. 121-138, 1993.
+    .. [1] C. Gueymard, "Critical analysis and performance assessment of
+       clear sky solar irradiance models using theoretical and measured
+       data," Solar Energy, vol. 51, pp. 121-138, 1993.
     '''
 
     airmass_absolute = airmass_relative * pressure / 101325.
@@ -136,32 +135,25 @@ def get_absolute_airmass(airmass_relative, pressure=101325.):
 
 def get_relative_airmass(zenith, model='kastenyoung1989'):
     '''
-    Gives the relative (not pressure-corrected) airmass.
+    Calculate relative (not pressure-adjusted) airmass at sea level.
 
-    Gives the airmass at sea-level when given a sun zenith angle (in
-    degrees). The ``model`` variable allows selection of different
-    airmass models (described below). If ``model`` is not included or is
-    not valid, the default model is 'kastenyoung1989'.
+    Parameter ``model`` allows selection of different airmass models.
 
     Parameters
     ----------
     zenith : numeric
-        Zenith angle of the sun in degrees. Note that some models use
-        the apparent (refraction corrected) zenith angle, and some
-        models use the true (not refraction-corrected) zenith angle. See
-        model descriptions to determine which type of zenith angle is
-        required. Apparent zenith angles must be calculated at sea level.
+        Zenith angle of the sun. [degrees]
 
     model : string, default 'kastenyoung1989'
         Available models include the following:
 
         * 'simple' - secant(apparent zenith angle) -
-          Note that this gives -inf at zenith=90
+          Note that this gives -Inf at zenith=90
         * 'kasten1966' - See reference [1] -
           requires apparent sun zenith
         * 'youngirvine1967' - See reference [2] -
           requires true sun zenith
-        * 'kastenyoung1989' - See reference [3] -
+        * 'kastenyoung1989' (default) - See reference [3] -
           requires apparent sun zenith
         * 'gueymard1993' - See reference [4] -
           requires apparent sun zenith
@@ -173,34 +165,40 @@ def get_relative_airmass(zenith, model='kastenyoung1989'):
     Returns
     -------
     airmass_relative : numeric
-        Relative airmass at sea level. Will return NaN values for any
-        zenith angle greater than 90 degrees.
+        Relative airmass at sea level. Returns NaN values for any
+        zenith angle greater than 90 degrees. [unitless]
+
+    Notes
+    -----
+    Some models use apparent (refraction-adjusted) zenith angle while
+    other models use true (not refraction-adjusted) zenith angle. Apparent
+    zenith angles should be calculated at sea level.
 
     References
     ----------
-    [1] Fritz Kasten. "A New Table and Approximation Formula for the
-    Relative Optical Air Mass". Technical Report 136, Hanover, N.H.:
-    U.S. Army Material Command, CRREL.
+    .. [1] Fritz Kasten. "A New Table and Approximation Formula for the
+       Relative Optical Air Mass". Technical Report 136, Hanover, N.H.:
+       U.S. Army Material Command, CRREL.
 
-    [2] A. T. Young and W. M. Irvine, "Multicolor Photoelectric
-    Photometry of the Brighter Planets," The Astronomical Journal, vol.
-    72, pp. 945-950, 1967.
+    .. [2] A. T. Young and W. M. Irvine, "Multicolor Photoelectric
+       Photometry of the Brighter Planets," The Astronomical Journal, vol.
+       72, pp. 945-950, 1967.
 
-    [3] Fritz Kasten and Andrew Young. "Revised optical air mass tables
-    and approximation formula". Applied Optics 28:4735-4738
+    .. [3] Fritz Kasten and Andrew Young. "Revised optical air mass tables
+       and approximation formula". Applied Optics 28:4735-4738
 
-    [4] C. Gueymard, "Critical analysis and performance assessment of
-    clear sky solar irradiance models using theoretical and measured
-    data," Solar Energy, vol. 51, pp. 121-138, 1993.
+    .. [4] C. Gueymard, "Critical analysis and performance assessment of
+       clear sky solar irradiance models using theoretical and measured
+       data," Solar Energy, vol. 51, pp. 121-138, 1993.
 
-    [5] A. T. Young, "AIR-MASS AND REFRACTION," Applied Optics, vol. 33,
-    pp. 1108-1110, Feb 1994.
+    .. [5] A. T. Young, "AIR-MASS AND REFRACTION," Applied Optics, vol. 33,
+       pp. 1108-1110, Feb 1994.
 
-    [6] Keith A. Pickering. "The Ancient Star Catalog". DIO 12:1, 20,
+    .. [6] Keith A. Pickering. "The Ancient Star Catalog". DIO 12:1, 20,
 
-    [7] Matthew J. Reno, Clifford W. Hansen and Joshua S. Stein, "Global
-    Horizontal Irradiance Clear Sky Models: Implementation and Analysis"
-    Sandia Report, (2012).
+    .. [7] Matthew J. Reno, Clifford W. Hansen and Joshua S. Stein, "Global
+       Horizontal Irradiance Clear Sky Models: Implementation and Analysis"
+       Sandia Report, (2012).
     '''
 
     # set zenith values greater than 90 to nans
@@ -211,7 +209,7 @@ def get_relative_airmass(zenith, model='kastenyoung1989'):
 
     if 'kastenyoung1989' == model:
         am = (1.0 / (np.cos(zenith_rad) +
-              0.50572*(((6.07995 + (90 - z)) ** - 1.6364))))
+              0.50572*((6.07995 + (90 - z)) ** - 1.6364)))
     elif 'kasten1966' == model:
         am = 1.0 / (np.cos(zenith_rad) + 0.15*((93.885 - z) ** - 1.253))
     elif 'simple' == model:
@@ -251,7 +249,7 @@ def gueymard94_pw(temp_air, relative_humidity):
 
     .. math::
 
-           w = 0.1 H_v \rho_v
+           Pw = 0.1 H_v \rho_v
 
     using Eq. 2 in [2]_
 
@@ -259,33 +257,36 @@ def gueymard94_pw(temp_air, relative_humidity):
 
            \rho_v = 216.7 R_H e_s /T
 
-    :math:`H_v` is the apparant water vapor scale height (km). The
-    expression for :math:`H_v` is Eq. 4 in [2]_:
+    :math:`Pw` is the precipitable water (cm), :math:`H_v` is the apparent
+    water vapor scale height (km) and :math:`\rho_v` is the surface water
+    vapor density (g/m^3). . The expression for :math:`H_v` is Eq. 4 in [2]_:
 
     .. math::
 
-           H_v = 0.4976 + 1.5265*T/273.15 + \exp(13.6897*T/273.15 - 14.9188*(T/273.15)^3)
+           H_v = 0.4976 + 1.5265 \frac{T}{273.15}
+               + \exp \left(13.6897 \frac{T}{273.15}
+               - 14.9188 \left( \frac{T}{273.15} \right)^3 \right)
 
-    :math:`\rho_v` is the surface water vapor density (g/m^3). In the
-    expression :math:`\rho_v`, :math:`e_s` is the saturation water vapor
-    pressure (millibar). The
-    expression for :math:`e_s` is Eq. 1 in [3]_
+    In the expression for :math:`\rho_v`, :math:`e_s` is the saturation water
+    vapor pressure (millibar). The expression for :math:`e_s` is Eq. 1 in [3]_
 
     .. math::
 
-          e_s = \exp(22.330 - 49.140*(100/T) - 10.922*(100/T)^2 - 0.39015*T/100)
+          e_s = \exp \left(22.330 - 49.140 \frac{100}{T} -
+              10.922 \left(\frac{100}{T}\right)^2 -
+              0.39015 \frac{T}{100} \right)
 
     Parameters
     ----------
     temp_air : numeric
-        ambient air temperature at the surface (C)
+        ambient air temperature :math:`T` at the surface. [C]
     relative_humidity : numeric
-        relative humidity at the surface (%)
+        relative humidity :math:`R_H` at the surface. [%]
 
     Returns
     -------
     pw : numeric
-        precipitable water (cm)
+        precipitable water. [cm]
 
     References
     ----------
@@ -325,27 +326,27 @@ def first_solar_spectral_correction(pw, airmass_absolute,
                                     min_pw=0.1, max_pw=8):
     r"""
     Spectral mismatch modifier based on precipitable water and absolute
-    (pressure corrected) airmass.
+    (pressure-adjusted) airmass.
 
-    Estimates a spectral mismatch modifier M representing the effect on
+    Estimates a spectral mismatch modifier :math:`M` representing the effect on
     module short circuit current of variation in the spectral
-    irradiance. M is estimated from absolute (pressure currected) air
-    mass, AMa, and precipitable water, Pwat, using the following
+    irradiance. :math:`M`  is estimated from absolute (pressure currected) air
+    mass, :math:`AM_a`, and precipitable water, :math:`Pw`, using the following
     function:
 
     .. math::
 
-        M = c_1 + c_2*AMa  + c_3*Pwat  + c_4*AMa^.5
-            + c_5*Pwat^.5 + c_6*AMa/Pwat^.5
+        M = c_1 + c_2 AM_a  + c_3 Pw  + c_4 AM_a^{0.5}
+            + c_5 Pw^{0.5} + c_6 \frac{AM_a} {Pw^{0.5}}
 
     Default coefficients are determined for several cell types with
     known quantum efficiency curves, by using the Simple Model of the
     Atmospheric Radiative Transfer of Sunshine (SMARTS) [1]_. Using
     SMARTS, spectrums are simulated with all combinations of AMa and
-    Pwat where:
+    Pw where:
 
-       * 0.5 cm <= Pwat <= 5 cm
-       * 1.0 <= AMa <= 5.0
+       * :math:`0.5 \textrm{cm} <= Pw <= 5 \textrm{cm}`
+       * :math:`1.0 <= AM_a <= 5.0`
        * Spectral range is limited to that of CMP11 (280 nm to 2800 nm)
        * spectrum simulated on a plane normal to the sun
        * All other parameters fixed at G173 standard
@@ -355,51 +356,49 @@ def first_solar_spectral_correction(pw, airmass_absolute,
     applied to fit Eq. 1 to determine the coefficients for each module.
 
     Based on the PVLIB Matlab function ``pvl_FSspeccorr`` by Mitchell
-    Lee and Alex Panchula, at First Solar, 2016 [2]_.
+    Lee and Alex Panchula of First Solar, 2016 [2]_.
 
     Parameters
     ----------
     pw : array-like
-        atmospheric precipitable water (cm).
+        atmospheric precipitable water. [cm]
 
     airmass_absolute : array-like
-        absolute (pressure corrected) airmass.
+        absolute (pressure-adjusted) airmass. [unitless]
 
     min_pw : float, default 0.1
-        minimum atmospheric precipitable water (cm). A lower pw value will be
-        automatically set to this minimum value to avoid model divergence.
+        minimum atmospheric precipitable water. Any pw value lower than min_pw
+        is set to min_pw to avoid model divergence. [cm]
 
     max_pw : float, default 8
-        maximum atmospheric precipitable water (cm). If a higher value is
-        encountered it will be set to np.nan to avoid model divergence.
+        maximum atmospheric precipitable water. Any pw value higher than max_pw
+        is set to NaN to avoid model divergence. [cm]
 
     module_type : None or string, default None
-        a string specifying a cell type. Can be lower or upper case
-        letters. Admits values of 'cdte', 'monosi', 'xsi', 'multisi',
-        'polysi'. If provided, this input selects coefficients for the
-        following default modules:
+        a string specifying a cell type. Values of 'cdte', 'monosi', 'xsi',
+        'multisi', and 'polysi' (can be lower or upper case). If provided,
+        module_type selects default coefficients for the following modules:
 
-            * 'cdte' - First Solar Series 4-2 CdTe modules.
-            * 'monosi', 'xsi' - First Solar TetraSun modules.
-            * 'multisi', 'polysi' - multi-crystalline silicon modules.
-            * 'cigs' - anonymous copper indium gallium selenide PV module
-            * 'asi' - anonymous amorphous silicon PV module
+            * 'cdte' - First Solar Series 4-2 CdTe module.
+            * 'monosi', 'xsi' - First Solar TetraSun module.
+            * 'multisi', 'polysi' - anonymous multi-crystalline silicon module.
+            * 'cigs' - anonymous copper indium gallium selenide module.
+            * 'asi' - anonymous amorphous silicon module.
 
         The module used to calculate the spectral correction
-        coefficients corresponds to the Mult-crystalline silicon
-        Manufacturer 2 Model C from [3]_. Spectral Response (SR) of CIGS
+        coefficients corresponds to the Multi-crystalline silicon
+        Manufacturer 2 Model C from [3]_. The spectral response (SR) of CIGS
         and a-Si modules used to derive coefficients can be found in [4]_
 
     coefficients : None or array-like, default None
-        allows for entry of user defined spectral correction
+        Allows for entry of user-defined spectral correction
         coefficients. Coefficients must be of length 6. Derivation of
         coefficients requires use of SMARTS and PV module quantum
         efficiency curve. Useful for modeling PV module types which are
         not included as defaults, or to fine tune the spectral
-        correction to a particular mono-Si, multi-Si, or CdTe PV module.
-        Note that the parameters for modules with very similar QE should
-        be similar, in most cases limiting the need for module specific
-        coefficients.
+        correction to a particular PV module. Note that the parameters for
+        modules with very similar quantum efficiency should be similar,
+        in most cases limiting the need for module specific coefficients.
 
     Returns
     -------
@@ -421,24 +420,24 @@ def first_solar_spectral_correction(pw, airmass_absolute,
        Models for PV Module Performance. National Renewable Energy
        Laboratory, 2014. http://www.nrel.gov/docs/fy14osti/61610.pdf
     .. [4] Schweiger, M. and Hermann, W, Influence of Spectral Effects
-        on Energy Yield of Different PV Modules: Comparison of Pwat and
-        MMF Approach, TUV Rheinland Energy GmbH report 21237296.003,
-        January 2017
+       on Energy Yield of Different PV Modules: Comparison of Pwat and
+       MMF Approach, TUV Rheinland Energy GmbH report 21237296.003,
+       January 2017
     """
 
     # --- Screen Input Data ---
 
-    # *** Pwat ***
-    # Replace Pwat Values below 0.1 cm with 0.1 cm to prevent model from
+    # *** Pw ***
+    # Replace Pw Values below 0.1 cm with 0.1 cm to prevent model from
     # diverging"
     pw = np.atleast_1d(pw)
     pw = pw.astype('float64')
     if np.min(pw) < min_pw:
         pw = np.maximum(pw, min_pw)
-        warn('Exceptionally low pw values replaced with {0} cm to prevent '
-             'model divergence'.format(min_pw))
+        warn(f'Exceptionally low pw values replaced with {min_pw} cm to '
+             'prevent model divergence')
 
-    # Warn user about Pwat data that is exceptionally high
+    # Warn user about Pw data that is exceptionally high
     if np.max(pw) > max_pw:
         pw[pw > max_pw] = np.nan
         warn('Exceptionally high pw values replaced by np.nan: '
@@ -503,28 +502,28 @@ def bird_hulstrom80_aod_bb(aod380, aod500):
     Parameters
     ----------
     aod380 : numeric
-        AOD measured at 380 nm
+        AOD measured at 380 nm. [unitless]
     aod500 : numeric
-        AOD measured at 500 nm
+        AOD measured at 500 nm. [unitless]
 
     Returns
     -------
     aod_bb : numeric
-        broadband AOD
+        Broadband AOD.  [unitless]
 
     See also
     --------
-    kasten96_lt
+    pvlib.atmosphere.kasten96_lt
 
     References
     ----------
-    [1] Bird and Hulstrom, "Direct Insolation Models" (1980)
-    `SERI/TR-335-344 <http://www.nrel.gov/docs/legosti/old/344.pdf>`_
+    .. [1] Bird and Hulstrom, "Direct Insolation Models" (1980)
+       `SERI/TR-335-344 <http://www.nrel.gov/docs/legosti/old/344.pdf>`_
 
-    [2] R. E. Bird and R. L. Hulstrom, "Review, Evaluation, and Improvement of
-    Direct Irradiance Models", Journal of Solar Energy Engineering 103(3),
-    pp. 182-192 (1981)
-    :doi:`10.1115/1.3266239`
+    .. [2] R. E. Bird and R. L. Hulstrom, "Review, Evaluation, and Improvement
+       of Direct Irradiance Models", Journal of Solar Energy Engineering
+       103(3), pp. 182-192 (1981)
+       :doi:`10.1115/1.3266239`
     """
     # approximate broadband AOD using (Bird-Hulstrom 1980)
     return 0.27583 * aod380 + 0.35 * aod500
@@ -532,7 +531,7 @@ def bird_hulstrom80_aod_bb(aod380, aod500):
 
 def kasten96_lt(airmass_absolute, precipitable_water, aod_bb):
     """
-    Calculate Linke turbidity factor using Kasten pyrheliometric formula.
+    Calculate Linke turbidity  using Kasten pyrheliometric formula.
 
     Note that broadband aerosol optical depth (AOD) can be approximated by AOD
     measured at 700 nm according to Molineaux [4] . Bird and Hulstrom offer an
@@ -541,56 +540,56 @@ def kasten96_lt(airmass_absolute, precipitable_water, aod_bb):
     Based on original implementation by Armel Oumbe.
 
     .. warning::
-        These calculations are only valid for air mass less than 5 atm and
+        These calculations are only valid for airmass less than 5 and
         precipitable water less than 5 cm.
 
     Parameters
     ----------
     airmass_absolute : numeric
-        airmass, pressure corrected in atmospheres
+        Pressure-adjusted airmass. [unitless]
     precipitable_water : numeric
-        precipitable water or total column water vapor in centimeters
+        Precipitable water. [cm]
     aod_bb : numeric
-        broadband AOD
+        broadband AOD. [unitless]
 
     Returns
     -------
     lt : numeric
-        Linke turbidity
+        Linke turbidity. [unitless]
 
     See also
     --------
-    bird_hulstrom80_aod_bb
-    angstrom_aod_at_lambda
+    pvlib.atmosphere.bird_hulstrom80_aod_bb
+    pvlib.atmosphere.angstrom_aod_at_lambda
 
     References
     ----------
-    [1] F. Linke, "Transmissions-Koeffizient und Trubungsfaktor", Beitrage
-    zur Physik der Atmosphare, Vol 10, pp. 91-103 (1922)
+    .. [1] F. Linke, "Transmissions-Koeffizient und Trubungsfaktor", Beitrage
+       zur Physik der Atmosphare, Vol 10, pp. 91-103 (1922)
 
-    [2] F. Kasten, "A simple parameterization of the pyrheliometric formula for
-    determining the Linke turbidity factor", Meteorologische Rundschau 33,
-    pp. 124-127 (1980)
+    .. [2] F. Kasten, "A simple parameterization of the pyrheliometric formula
+       for determining the Linke turbidity factor", Meteorologische Rundschau
+       33, pp. 124-127 (1980)
 
-    [3] Kasten, "The Linke turbidity factor based on improved values of the
-    integral Rayleigh optical thickness", Solar Energy, Vol. 56, No. 3,
-    pp. 239-244 (1996)
-    :doi:`10.1016/0038-092X(95)00114-7`
+    .. [3] Kasten, "The Linke turbidity factor based on improved values of the
+       integral Rayleigh optical thickness", Solar Energy, Vol. 56, No. 3,
+       pp. 239-244 (1996)
+       :doi:`10.1016/0038-092X(95)00114-7`
 
-    [4] B. Molineaux, P. Ineichen, N. O'Neill, "Equivalence of pyrheliometric
-    and monochromatic aerosol optical depths at a single key wavelength",
-    Applied Optics Vol. 37, issue 10, 7008-7018 (1998)
-    :doi:`10.1364/AO.37.007008`
+    .. [4] B. Molineaux, P. Ineichen, N. O'Neill, "Equivalence of
+       pyrheliometric and monochromatic aerosol optical depths at a single key
+       wavelength", Applied Optics Vol. 37, issue 10, 7008-7018 (1998)
+       :doi:`10.1364/AO.37.007008`
 
-    [5] P. Ineichen, "Conversion function between the Linke turbidity and the
-    atmospheric water vapor and aerosol content", Solar Energy 82,
-    pp. 1095-1097 (2008)
-    :doi:`10.1016/j.solener.2008.04.010`
+    .. [5] P. Ineichen, "Conversion function between the Linke turbidity and
+       the atmospheric water vapor and aerosol content", Solar Energy 82,
+       pp. 1095-1097 (2008)
+       :doi:`10.1016/j.solener.2008.04.010`
 
-    [6] P. Ineichen and R. Perez, "A new airmass independent formulation for
-    the Linke Turbidity coefficient", Solar Energy, Vol. 73, no. 3, pp. 151-157
-    (2002)
-    :doi:`10.1016/S0038-092X(02)00045-2`
+    .. [6] P. Ineichen and R. Perez, "A new airmass independent formulation for
+       the Linke Turbidity coefficient", Solar Energy, Vol. 73, no. 3,
+       pp. 151-157 (2002)
+       :doi:`10.1016/S0038-092X(02)00045-2`
     """
     # "From numerically integrated spectral simulations done with Modtran
     # (Berk, 1989), Molineaux (1998) obtained for the broadband optical depth
@@ -625,33 +624,35 @@ def angstrom_aod_at_lambda(aod0, lambda0, alpha=1.14, lambda1=700.0):
     Parameters
     ----------
     aod0 : numeric
-        aerosol optical depth (AOD) measured at known wavelength
+        Aerosol optical depth (AOD) measured at wavelength ``lambda0``.
+        [unitless]
     lambda0 : numeric
-        wavelength in nanometers corresponding to ``aod0``
+        Wavelength corresponding to ``aod0``. [nm]
     alpha : numeric, default 1.14
-        Angstrom :math:`\alpha` exponent corresponding to ``aod0``
+        Angstrom :math:`\alpha` exponent corresponding to ``aod0``. [unitless]
     lambda1 : numeric, default 700
-        desired wavelength in nanometers
+        Desired wavelength. [nm]
 
     Returns
     -------
     aod1 : numeric
-        AOD at desired wavelength, ``lambda1``
+        AOD at desired wavelength ``lambda1``. [unitless]
 
     See also
     --------
-    angstrom_alpha
+    pvlib.atmosphere.angstrom_alpha
 
     References
     ----------
-    [1] Anders Angstrom, "On the Atmospheric Transmission of Sun Radiation and
-    On Dust in the Air", Geografiska Annaler Vol. 11, pp. 156-166 (1929) JSTOR
-    :doi:`10.2307/519399`
+    .. [1] Anders Angstrom, "On the Atmospheric Transmission of Sun Radiation
+       and On Dust in the Air", Geografiska Annaler Vol. 11, pp. 156-166 (1929)
+       JSTOR
+       :doi:`10.2307/519399`
 
-    [2] Anders Angstrom, "Techniques of Determining the Turbidity of the
-    Atmosphere", Tellus 13:2, pp. 214-223 (1961) Taylor & Francis
-    :doi:`10.3402/tellusa.v13i2.9493` and Co-Action Publishing
-    :doi:`10.1111/j.2153-3490.1961.tb00078.x`
+    .. [2] Anders Angstrom, "Techniques of Determining the Turbidity of the
+       Atmosphere", Tellus 13:2, pp. 214-223 (1961) Taylor & Francis
+       :doi:`10.3402/tellusa.v13i2.9493` and Co-Action Publishing
+       :doi:`10.1111/j.2153-3490.1961.tb00078.x`
     """
     return aod0 * ((lambda1 / lambda0) ** (-alpha))
 
@@ -663,21 +664,22 @@ def angstrom_alpha(aod1, lambda1, aod2, lambda2):
     Parameters
     ----------
     aod1 : numeric
-        first aerosol optical depth
+        Aerosol optical depth at wavelength ``lambda1``. [unitless]
     lambda1 : numeric
-        wavelength in nanometers corresponding to ``aod1``
+        Wavelength corresponding to ``aod1``. [nm]
     aod2 : numeric
-        second aerosol optical depth
+        Aerosol optical depth  at wavelength ``lambda2``. [unitless]
     lambda2 : numeric
-        wavelength in nanometers corresponding to ``aod2``
+        Wavelength corresponding to ``aod2``. [nm]
 
     Returns
     -------
     alpha : numeric
-        Angstrom :math:`\alpha` exponent for AOD in ``(lambda1, lambda2)``
+        Angstrom :math:`\alpha` exponent for wavelength in
+        ``(lambda1, lambda2)``. [unitless]
 
     See also
     --------
-    angstrom_aod_at_lambda
+    pvlib.atmosphere.angstrom_aod_at_lambda
     """
     return - np.log(aod1 / aod2) / np.log(lambda1 / lambda2)
