@@ -1,6 +1,5 @@
-from pvlib.iotools.tmy import read_tmy2  # noqa: F401
-from pvlib.iotools.tmy import read_tmy3  # noqa: F401
-from pvlib.iotools.epw import read_epw  # noqa: F401
+from pvlib.iotools.tmy import read_tmy2, read_tmy3  # noqa: F401
+from pvlib.iotools.epw import read_epw, parse_epw  # noqa: F401
 from pvlib.iotools.srml import read_srml  # noqa: F401
 from pvlib.iotools.srml import read_srml_month_from_solardat  # noqa: F401
 from pvlib.iotools.surfrad import read_surfrad  # noqa: F401
@@ -11,3 +10,14 @@ from pvlib.iotools.ecmwf_macc import get_ecmwf_macc  # noqa: F401
 from pvlib.iotools.crn import read_crn  # noqa: F401
 from pvlib.iotools.solrad import read_solrad  # noqa: F401
 from pvlib.iotools.psm3 import get_psm3  # noqa: F401
+from pvlib.iotools.psm3 import read_psm3  # noqa: F401
+from pvlib.iotools.psm3 import parse_psm3  # noqa: F401
+from pvlib.iotools.pvgis import get_pvgis_tmy, read_pvgis_tmy  # noqa: F401
+from pvlib.iotools.pvgis import read_pvgis_hourly  # noqa: F401
+from pvlib.iotools.pvgis import get_pvgis_hourly  # noqa: F401
+from pvlib.iotools.bsrn import get_bsrn  # noqa: F401
+from pvlib.iotools.bsrn import read_bsrn  # noqa: F401
+from pvlib.iotools.bsrn import parse_bsrn  # noqa: F401
+from pvlib.iotools.sodapro import get_cams  # noqa: F401
+from pvlib.iotools.sodapro import read_cams  # noqa: F401
+from pvlib.iotools.sodapro import parse_cams  # noqa: F401
